requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
Django==4.2.15
gunicorn==22.0.0
h11==0.14.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
whitenoise==6.7.0
